Change Default RDP Version to v10
Problem I'm solving:
When importing RDP files, mRemoteNG defaults to RDPv6 (which is extremely old, and particularly does not support Gateway access tokens). Thus I would have to manually change the config to a newer RDP protocol, so that I could initiate  the connection.

It is safe to assume that the majority of Windows PCs on the internet today are running Windows 10 or above, and by keeping the default at RDP6 we are causing more unnecessary friction to those majority users.

I've changed the default to Rdp10 (released in Windows 10).
https://en.wikipedia.org/wiki/Remote_Desktop_Protocol#Version_10.0